requirements.txt
fastapi
uvicorn
requests
openai>=1.37.0
pydub==0.24.1
ffmpeg-python
python-telegram-bot==20.7
aiohttp
httpx